solved an issue with the f1data getting throttle only, added mguk deplyoment *intelligence* still needs some improvements, will be working on getting the Z of the track next

+++ include/physics.cuh
@@ -50,7 +50,7 @@ struct TrackSegment {
     float y;                            // y of segment
     float real_speed_kmh;               // Real speed of the segment (track data)
     float real_rpm;                     // Real RPM of the segment (track data)
-    float real_gear;                    // Real gear of the segment (track data)
+    int real_gear;                    // Real gear of the segment (track data)
     float real_throttle_pedal;          // Real value of throttle pedal
     float real_brake_pedal;             // Real value of brake pedal
 };
